feat(stats): split statistics page into four deep-linked sub-pages
Break the monolithic Statistics dashboard into Summary, Rules, Sessions
and Storage sub-pages, each with its own deeplink sub-URL
(#stats, #stats/rules, #stats/sessions, #stats/storage), a Radix TabNav
to switch between them, and the active sub-page reflected in the header
breadcrumb (same pattern as archived sessions).

- Summary shows the headline KPIs (grouped tabs, deduplicated tabs,
  pinned/active/archived session counts, total local storage size) plus
  the weekly grouping/deduplication trends.
- Rules, Sessions and Storage expose the detailed breakdowns; the reset
  button lives on the Rules sub-page.
- Extract shared statistics primitives and per-sub-page view components
  under src/components/Core/Statistics/.
- Make the lazy-load skeleton sub-tab aware.
- Add stats sub-route parsing to useDeepLinking with tests, new i18n
  keys (en/fr/es), updated stories and the statistics doc scenario.
