+++ e2e-doc-scenarios/scenarios/24-statistics.routing.ts
@@ -5,18 +5,23 @@ import type { Manifest } from '../../e2e-shared/routing/types.js';
 
 const ALL_LOCALES = ['en', 'fr', 'es'] as const;
 
+const starlightRoute = (capture: string, path: string) => ({
+  capture,
+  destinations: [
+    {
+      target: 'starlight' as const,
+      path,
+      locales: [...ALL_LOCALES],
+      themes: ['light', 'dark'] as const,
+    },
+  ],
+});
+
 export const STATISTICS_MANIFEST: Manifest = {
   routes: [
-    {
-      capture: 'statistics-overview',
-      destinations: [
-        {
-          target: 'starlight',
-          path: 'statistics-overview',
-          locales: [...ALL_LOCALES],
-          themes: ['light', 'dark'],
-        },
-      ],
-    },
+    starlightRoute('statistics-overview', 'statistics-overview'),
+    starlightRoute('statistics-rules', 'statistics-rules'),
+    starlightRoute('statistics-sessions', 'statistics-sessions'),
+    starlightRoute('statistics-storage', 'statistics-storage'),
   ],
 };
