+++ e2e-doc-scenarios/scenarios/24-statistics.scenario.ts
@@ -41,10 +41,26 @@ test('statistics overview', async (
     theme,
   );
   await page.getByTestId('page-stats').waitFor({ state: 'visible' });
+  await page.getByTestId('page-stats-summary').waitFor({ state: 'visible' });
   await page.waitForTimeout(300);
   await captureStep(page, 'statistics-overview', {
-    description: 'Statistics page with totals, weekly trend and top-rules bars.',
+    description: 'Statistics summary sub-page with headline KPIs and the weekly trend.',
   });
+
+  // Walk the four sub-tabs so the docs show each detailed view.
+  const subTabs: Array<{ tab: string; view: string; capture: string; description: string }> = [
+    { tab: 'page-stats-tab-rules', view: 'page-stats-rules', capture: 'statistics-rules', description: 'Statistics rules sub-page: lifetime totals, weekly trend and most active rules.' },
+    { tab: 'page-stats-tab-sessions', view: 'page-stats-sessions', capture: 'statistics-sessions', description: 'Statistics sessions sub-page: volumes, activity and top categories/domains.' },
+    { tab: 'page-stats-tab-storage', view: 'page-stats-storage', capture: 'statistics-storage', description: 'Statistics storage sub-page: quota usage and per-category breakdown.' },
+  ];
+
+  for (const { tab, view, capture, description } of subTabs) {
+    await page.getByTestId(tab).click();
+    await page.getByTestId(view).waitFor({ state: 'visible' });
+    await page.waitForTimeout(300);
+    await captureStep(page, capture, { description });
+  }
+
   await page.close();
 
   await writeScenarioReadme({
