+++ public/_locales/en/messages.json
@@ -555,6 +555,14 @@
   "domainRulesPageDescription": { "message": "Create rules that automatically group and deduplicate tabs based on their domain or URL pattern." },
   "sessionsPageDescription": { "message": "Capture your open tabs as sessions you can restore later. Pin a session to access it in one click from the popup." },
   "statisticsPageDescription": { "message": "Track how many tab groups have been created and how many duplicate tabs have been closed automatically." },
+  "statsSummaryTab": { "message": "Summary" },
+  "statsRulesTab": { "message": "Rules" },
+  "statsSessionsTab": { "message": "Sessions" },
+  "statsStorageTab": { "message": "Storage" },
+  "statsSummaryStorageTotal": { "message": "Local storage" },
+  "statsRulesPageDescription": { "message": "Detailed grouping and deduplication counters, including the most active rules." },
+  "statsSessionsPageDescription": { "message": "Detailed session metrics: volumes, activity, top categories and domains." },
+  "statsStoragePageDescription": { "message": "Detailed local storage usage, broken down by category." },
   "importExportPageDescription": { "message": "Back up or restore your rules and sessions as JSON files, with conflict resolution on import." },
   "settingsPageDescription": { "message": "Configure the extension's global preferences and behavior." },
   "sessionsEmptyStateTitle": { "message": "No saved sessions." },
