+++ public/_locales/es/messages.json
@@ -556,6 +556,14 @@
   "domainRulesPageDescription": { "message": "Crea reglas que agrupan y deduplican automáticamente las pestañas según su dominio o un patrón de URL." },
   "sessionsPageDescription": { "message": "Captura tus pestañas abiertas como sesiones que podrás restaurar más tarde. Fija una sesión para acceder a ella con un clic desde el popup." },
   "statisticsPageDescription": { "message": "Consulta cuántos grupos de pestañas se han creado y cuántas pestañas duplicadas se han cerrado automáticamente." },
+  "statsSummaryTab": { "message": "Resumen" },
+  "statsRulesTab": { "message": "Reglas" },
+  "statsSessionsTab": { "message": "Sesiones" },
+  "statsStorageTab": { "message": "Almacenamiento" },
+  "statsSummaryStorageTotal": { "message": "Almacenamiento local" },
+  "statsRulesPageDescription": { "message": "Detalle de los contadores de agrupación y deduplicación, con las reglas más activas." },
+  "statsSessionsPageDescription": { "message": "Detalle de las métricas de sesiones: volúmenes, actividad, categorías y dominios principales." },
+  "statsStoragePageDescription": { "message": "Detalle del uso del almacenamiento local, desglosado por categoría." },
   "importExportPageDescription": { "message": "Haz copia de seguridad o restaura tus reglas y sesiones en formato JSON, con resolución de conflictos al importar." },
   "settingsPageDescription": { "message": "Configura las preferencias globales y el comportamiento de la extensión." },
   "sessionsEmptyStateTitle": { "message": "No hay sesiones guardadas." },
