+++ public/_locales/fr/messages.json
@@ -556,6 +556,14 @@
   "domainRulesPageDescription": { "message": "Créez des règles qui regroupent et dédupliquent automatiquement les onglets selon leur domaine ou un motif d'URL." },
   "sessionsPageDescription": { "message": "Capturez vos onglets ouverts sous forme de sessions à restaurer plus tard. Épinglez une session pour y accéder en un clic depuis le popup." },
   "statisticsPageDescription": { "message": "Suivez le nombre de groupes d'onglets créés et celui des onglets dupliqués fermés automatiquement." },
+  "statsSummaryTab": { "message": "Résumé" },
+  "statsRulesTab": { "message": "Règles" },
+  "statsSessionsTab": { "message": "Sessions" },
+  "statsStorageTab": { "message": "Stockage" },
+  "statsSummaryStorageTotal": { "message": "Stockage local" },
+  "statsRulesPageDescription": { "message": "Détail des compteurs de regroupement et de déduplication, avec les règles les plus actives." },
+  "statsSessionsPageDescription": { "message": "Détail des métriques de sessions : volumes, activité, catégories et domaines principaux." },
+  "statsStoragePageDescription": { "message": "Détail de l'utilisation du stockage local, réparti par catégorie." },
   "importExportPageDescription": { "message": "Sauvegardez ou restaurez vos règles et vos sessions au format JSON, avec résolution des conflits à l'import." },
   "settingsPageDescription": { "message": "Configurez les préférences globales et le comportement de l'extension." },
   "sessionsEmptyStateTitle": { "message": "Aucune session sauvegardée." },
