+++ src/pages/skeletons/StatisticsPageSkeleton.tsx
@@ -1,33 +1,95 @@
 import { Box, Card, Flex, Grid, Skeleton } from '@radix-ui/themes';
 import { PageLayoutSkeleton } from '@/components/UI/PageLayout';
+import type { StatsSubTab } from '@/hooks/useDeepLinking';
 
-const CARDS: Array<{ heading: string; body: number }> = [
-  { heading: '160px', body: 120 },
-  { heading: '160px', body: 180 },
-  { heading: '140px', body: 100 },
-  { heading: '140px', body: 100 },
-  { heading: '120px', body: 150 },
-  { heading: '160px', body: 200 },
-  { heading: '150px', body: 160 },
-];
+interface StatisticsPageSkeletonProps {
+  statsTab?: StatsSubTab;
+}
 
-export function StatisticsPageSkeleton() {
+/** Faux TabNav row matching the four statistics sub-tabs. */
+function TabsRowSkeleton() {
+  return (
+    <Flex gap="4" mb="4" pb="2" style={{ borderBottom: '1px solid var(--gray-a4)' }}>
+      <Skeleton width="70px" height="20px" />
+      <Skeleton width="60px" height="20px" />
+      <Skeleton width="80px" height="20px" />
+      <Skeleton width="80px" height="20px" />
+    </Flex>
+  );
+}
+
+function CardSkeleton({ heading, body }: { heading: string; body: number }) {
+  return (
+    <Card>
+      <Flex direction="column" gap="3" p="2">
+        <Skeleton width={heading} height="20px" />
+        <Skeleton height={`${body}px`} />
+      </Flex>
+    </Card>
+  );
+}
+
+function SummarySkeleton() {
+  return (
+    <Box>
+      <Grid columns={{ initial: '2', sm: '3' }} gap="3">
+        {Array.from({ length: 6 }).map((_, i) => (
+          <Card key={i}>
+            <Flex direction="column" gap="2" p="3">
+              <Skeleton width="90px" height="16px" />
+              <Skeleton width="60px" height="32px" />
+            </Flex>
+          </Card>
+        ))}
+      </Grid>
+      <Box mt="4">
+        <CardSkeleton heading="120px" body={90} />
+      </Box>
+    </Box>
+  );
+}
+
+function RulesSkeleton() {
+  return (
+    <Box>
+      <Grid columns={{ initial: '1', md: '2' }} gap="4">
+        <CardSkeleton heading="160px" body={120} />
+        <CardSkeleton heading="120px" body={120} />
+      </Grid>
+      <Box mt="4">
+        <CardSkeleton heading="150px" body={160} />
+      </Box>
+      <Box mt="4">
+        <Skeleton width="120px" height="32px" />
+      </Box>
+    </Box>
+  );
+}
+
+function SessionsSkeleton() {
+  return (
+    <Grid columns={{ initial: '1', md: '2' }} gap="4">
+      <CardSkeleton heading="160px" body={180} />
+      <CardSkeleton heading="140px" body={140} />
+      <CardSkeleton heading="150px" body={200} />
+      <CardSkeleton heading="160px" body={200} />
+    </Grid>
+  );
+}
+
+function StorageSkeleton() {
+  return <CardSkeleton heading="160px" body={240} />;
+}
+
+export function StatisticsPageSkeleton({ statsTab = 'summary' }: StatisticsPageSkeletonProps) {
   return (
     <PageLayoutSkeleton titleKey="statisticsTab" descriptionKey="statisticsPageDescription">
       <Box data-testid="page-stats-skeleton">
-        <Grid columns={{ initial: '1', md: '2' }} gap="4">
-          {CARDS.map((card, i) => (
-            <Card key={i}>
-              <Flex direction="column" gap="3" p="2">
-                <Skeleton width={card.heading} height="20px" />
-                <Skeleton height={`${card.body}px`} />
-              </Flex>
-            </Card>
-          ))}
-        </Grid>
-        <Box mt="4">
-          <Skeleton width="120px" height="32px" />
-        </Box>
+        <TabsRowSkeleton />
+        {statsTab === 'summary' && <SummarySkeleton />}
+        {statsTab === 'rules' && <RulesSkeleton />}
+        {statsTab === 'sessions' && <SessionsSkeleton />}
+        {statsTab === 'storage' && <StorageSkeleton />}
       </Box>
     </PageLayoutSkeleton>
   );
